pyproject.toml
[project]
name = "llm-based-rock-paper-scissors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.5",
    "openai>=2.6.1",
    "sentencepiece>=0.2.1",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
